삭제하기전에 confirm 메세지, 삭제 후 리스트링크 노출

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -19,7 +19,7 @@
            <li class="flex">
                <a class="w-[40px]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%></a>
                <a class="flex-grow hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
-               <a class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
+               <a onclick="if( !confirm('정말로 삭제하시겠습니까?'))return false;" class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
            </li>
            <% } %>
        </ul>
